requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
pytz>=2024.1
web3==6.11.1
aiohttp==3.8.5
boto3==1.28.44
python-dateutil==2.8.2